+++ showcase/integrations/google-adk/src/app/demos/declarative-hashbrown/charts/bar-chart.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 /**
- * BarChart ported from showcase/starters/template/frontend/components/charts/bar-chart.tsx
- * for the byoc-hashbrown demo (Wave 4a).
+ * BarChart ported from showcase/shared/starter-template/components/charts/bar-chart.tsx
+ * for the declarative-hashbrown demo.
  *
  * Adds data-testid="bar-chart" on the top-level container for E2E coverage.
  */
