+++ showcase/integrations/google-adk/src/app/demos/declarative-hashbrown/charts/chart-config.ts
@@ -1,8 +1,8 @@
 /**
  * CopilotKit brand chart palette -- Plus Jakarta Sans / brand color system.
  *
- * Ported verbatim from showcase/starters/template/frontend/components/charts/chart-config.ts
- * for the byoc-hashbrown demo (Wave 4a).
+ * Ported verbatim from showcase/shared/starter-template/components/charts/chart-config.ts
+ * for the declarative-hashbrown demo.
  */
 export const CHART_COLORS = [
   "#BEC2FF", // lilac-400
