+++ showcase/integrations/google-adk/src/app/demos/declarative-hashbrown/charts/pie-chart.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 /**
- * PieChart ported from showcase/starters/template/frontend/components/charts/pie-chart.tsx
- * for the byoc-hashbrown demo (Wave 4a).
+ * PieChart ported from showcase/shared/starter-template/components/charts/pie-chart.tsx
+ * for the declarative-hashbrown demo.
  *
  * Adds data-testid="pie-chart" on the top-level container for E2E coverage.
  */
