+++ showcase/integrations/google-adk/src/app/demos/declarative-hashbrown/chat.tsx
@@ -1,0 +1,24 @@
+"use client";
+
+import {
+  CopilotChat,
+  CopilotChatAssistantMessage,
+} from "@copilotkit/react-core/v2";
+import { HashBrownRenderMessage } from "./hashbrown-renderer";
+import { useByocHashbrownSuggestions } from "./suggestions";
+
+export function Chat() {
+  useByocHashbrownSuggestions();
+
+  return (
+    <CopilotChat
+      className="h-full"
+      messageView={{
+        // The renderer reads only `message` from the slot props; cast to the
+        // wider CopilotChatAssistantMessage signature to satisfy the slot type.
+        assistantMessage:
+          HashBrownRenderMessage as unknown as typeof CopilotChatAssistantMessage,
+      }}
+    />
+  );
+}
