+++ showcase/integrations/google-adk/src/app/demos/declarative-hashbrown/metric-card.tsx
@@ -1,13 +1,5 @@
 "use client";
 
-/**
- * MetricCard for the byoc-hashbrown demo (Wave 4a).
- *
- * Extracted from showcase/starters/template/frontend/components/renderers/hashbrown/index.tsx
- * so the hashbrown-renderer can import it as its own module.
- */
-import React from "react";
-
 interface MetricCardProps {
   label: string;
   value: string;
