+++ showcase/integrations/google-adk/src/app/demos/declarative-hashbrown/types.ts
@@ -1,0 +1,10 @@
+export const SALES_STAGES = [
+  "prospect",
+  "qualified",
+  "proposal",
+  "negotiation",
+  "closed-won",
+  "closed-lost",
+] as const;
+
+export type SalesStage = (typeof SALES_STAGES)[number];
